+++ scripts/newtab.md.sh
@@ -16,10 +16,10 @@ sed "1,/REPLACETHIS/ d" newtab.template.html >> "$newtabtemp"
 sed "/REPLACE_ME_WITH_THE_CHANGE_LOG_USING_SED/,$ d" "$newtabtemp" > "$newtab"
 echo """
 <input type="checkbox"  id="spoilerbutton" />
-<label for="spoilerbutton" onclick="">Changelogs</label>
+<label for="spoilerbutton" onclick="">Changelog</label>
 <div class="spoiler">
 """ >> "$newtab"
-marked ../../doc/changelog.md >> "$newtab"
+marked ../../CHANGELOG.md >> "$newtab"
 echo """
 </div>
 """ >> "$newtab"